+++ core/src/utils/ProcessingConfig.ts
@@ -1,0 +1,106 @@
+//  重置 create-react-app 中的 react-script配置
+import { reactScripts, webpackConfigPath } from "../overrides/pathUtils"
+import webpackNodeExternals from "webpack-node-externals"
+import webpack from "webpack"
+import { OptionsProps } from "../interface"
+import { OverridesProps } from "../overrides"
+import CreateTemporaryAsset from "./plugins/CreateTemporaryAsset"
+import { restOutPut } from "./output"
+import { getWbpackBarPlugins, restWebpackManifestPlugin, clearHtmlTemp, addServerPlugins, AddDefinePlugin } from "./plugins"
+
+export type GetWebpackConfig = (
+  newConfig: webpack.Configuration,
+  type: "server" | "client",
+  overrides: OverridesProps,
+  nodeExternals: boolean,
+  split: boolean,
+  env: "development" | "production",
+  isWebpackDevServer: boolean,
+  options: OptionsProps
+) => webpack.Configuration
+
+
+const getWebpackConfig: GetWebpackConfig = (newConfig, type, overrides, nodeExternals, split, env, isWebpackDevServer, options) => {
+  /** 入口 */
+  newConfig.entry = overrides[`${type}_path`]
+  /** 加载 进度条 plugin */
+  newConfig = getWbpackBarPlugins(newConfig, {
+    name: type,
+  })
+  /** 输出配置 */
+  const out: webpack.Configuration["output"] = {
+    filename: `${type}.js`,
+    path: overrides.output_path,
+  }
+  if (type === "server") {
+    /** 输出 类型 */
+    out.library = { type: "commonjs2" }
+  }
+
+  const HOST = process.env.HOST || 'localhost'
+  const PORT = process.env.PORT || 3000
+
+  const httpPath = `http://${HOST}:${PORT}`
+
+  /** start 命令时候 配置前缀 为 devServer 端口  */
+  if (isWebpackDevServer && env === "development") {
+    out.publicPath = `${httpPath}/`
+  } else {
+    out.publicPath = `/`
+  }
+
+  /** 重置 输入配置 */
+  newConfig = restOutPut(newConfig, out)
+
+  /** start 命令下 生成 client 端 asset 文件 **/
+  let isCreateAsset = false;
+
+  if (isWebpackDevServer && type === "client" && env === "development") {
+    isCreateAsset = true
+  }
+  /** start 命令下  生成 server.js文件和 自动启动 server.js 服务  */
+  if (isWebpackDevServer && type === "server" && env === "development") {
+    newConfig = addServerPlugins(newConfig, out)
+  }
+  /**  重置 asset-manifest.json 文件内容 */
+  newConfig = restWebpackManifestPlugin(newConfig, overrides.paths, type, isCreateAsset, httpPath)
+  /** 清除 html 模板方面的 plugin  **/
+  newConfig = clearHtmlTemp(newConfig)
+
+  newConfig.module.exprContextCritical = false;
+  // 添加 define plugin
+  newConfig = AddDefinePlugin(newConfig, overrides, isWebpackDevServer)
+
+  newConfig.plugins.push(
+    new CreateTemporaryAsset(`${overrides.output_path}/asset-${type}-manifest.json`)
+  )
+
+  if (!split) {
+    // 代码是否进行分割
+    newConfig.plugins.push(new webpack.optimize.LimitChunkCountPlugin({ maxChunks: 1 }))
+  }
+
+  if (nodeExternals) {
+    /** 
+     * https://www.npmjs.com/package/webpack-node-externals
+     * 这个库扫描node_modules文件夹中的所有 node_modules 名称，并构建一个外部函数，告诉 Webpack 不要捆绑这些模块或它们的任何子模块 
+     * */
+    newConfig.externals = [webpackNodeExternals()]
+  }
+
+  if (options.miniServer && type === "server") {
+    /** server 端 去除代码压缩 */
+    newConfig.optimization.minimize = false
+    newConfig.optimization.minimizer = []
+  }
+
+  if (options.miniClient && type === "client") {
+    /** client 端 去除代码压缩 */
+    newConfig.optimization.minimize = false
+    newConfig.optimization.minimizer = []
+  }
+
+  return newConfig
+}
+
+export default getWebpackConfig
